+++ Examples/Docs/Sources/DocsExamples/main.swift
@@ -1,0 +1,34 @@
+import Foundation
+import MongoSwift
+
+// swiftlint:disable force_unwrapping
+
+/// Examples used for the MongoDB documentation on Causal Consistency.
+/// - SeeAlso: https://docs.mongodb.com/manual/core/read-isolation-consistency-recency/#examples
+private func causalConsistency() throws {
+    // Start example 1
+    let client1 = try MongoClient()
+    let s1 = try client1.startSession(options: ClientSessionOptions(causalConsistency: true))
+    let currentDate = Date()
+    var dbOptions = DatabaseOptions(readConcern: ReadConcern(.majority),
+                                    writeConcern: try WriteConcern(w: .majority, wtimeoutMS: 1000))
+    let items = client1.db("test", options: dbOptions).collection("items")
+    try items.updateOne(filter: ["sku": "111", "end": BSONNull()],
+                        update: ["$set": ["end": currentDate] as Document],
+                        session: s1)
+    try items.insertOne(["sku": "nuts-111", "name": "Pecans", "start": currentDate], session: s1)
+
+    // Start example 2
+    let client2 = try MongoClient()
+    try client2.withSession(options: ClientSessionOptions(causalConsistency: true)) { s2 in
+        // The cluster and operation times are guaranteed to be non-nil since we already used s1 for operations above.
+        s2.advanceClusterTime(to: s1.clusterTime!)
+        s2.advanceOperationTime(to: s1.operationTime!)
+
+        dbOptions.readPreference = ReadPreference(.secondary)
+        let items2 = client2.db("test", options: dbOptions).collection("items")
+        for item in try items2.find(["end": BSONNull()], session: s2) {
+            print(item)
+        }
+    }
+}
